+++ src/main/java/com/meoguri/linkocean/domain/bookmark/entity/Bookmark.java
@@ -16,6 +16,7 @@ import javax.persistence.ManyToOne;
 
 import com.meoguri.linkocean.domain.BaseIdEntity;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType;
+import com.meoguri.linkocean.domain.linkmetadata.entity.LinkMetadata;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
 
 import lombok.Builder;
