+++ src/main/java/com/meoguri/linkocean/domain/linkmetadata/entity/LinkMetadata.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.bookmark.entity;
+package com.meoguri.linkocean.domain.linkmetadata.entity;
 
 import static lombok.AccessLevel.*;
 
@@ -7,7 +7,7 @@ import javax.persistence.Embedded;
 import javax.persistence.Entity;
 
 import com.meoguri.linkocean.domain.BaseIdEntity;
-import com.meoguri.linkocean.domain.bookmark.entity.vo.Url;
+import com.meoguri.linkocean.domain.linkmetadata.entity.vo.Url;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
