+++ src/main/java/com/meoguri/linkocean/domain/linkmetadata/repository/LinkMetadataRepository.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.bookmark.repository;
+package com.meoguri.linkocean.domain.linkmetadata.repository;
 
 import java.util.Optional;
 
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.meoguri.linkocean.domain.bookmark.entity.LinkMetadata;
-import com.meoguri.linkocean.domain.bookmark.entity.vo.Url;
+import com.meoguri.linkocean.domain.linkmetadata.entity.LinkMetadata;
+import com.meoguri.linkocean.domain.linkmetadata.entity.vo.Url;
 
 public interface LinkMetadataRepository extends JpaRepository<LinkMetadata, Long> {
