+++ src/test/java/com/meoguri/linkocean/domain/linkmetadata/entity/LinkMetadataTest.java
@@ -1,11 +1,11 @@
-package com.meoguri.linkocean.domain.bookmark.entity;
+package com.meoguri.linkocean.domain.linkmetadata.entity;
 
 import static com.meoguri.linkocean.domain.util.Fixture.*;
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import com.meoguri.linkocean.domain.bookmark.entity.vo.Url;
+import com.meoguri.linkocean.domain.linkmetadata.entity.vo.Url;
 
 class LinkMetadataTest {
