+++ src/test/java/com/meoguri/linkocean/domain/linkmetadata/entity/vo/UrlTest.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.bookmark.entity.vo;
+package com.meoguri.linkocean.domain.linkmetadata.entity.vo;
 
 import static org.assertj.core.api.Assertions.*;
