+++ src/test/java/com/meoguri/linkocean/domain/linkmetadata/repository/LinkMetadataRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.bookmark.repository;
+package com.meoguri.linkocean.domain.linkmetadata.repository;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -12,8 +12,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import com.meoguri.linkocean.domain.bookmark.entity.LinkMetadata;
-import com.meoguri.linkocean.domain.bookmark.entity.vo.Url;
+import com.meoguri.linkocean.domain.linkmetadata.entity.LinkMetadata;
+import com.meoguri.linkocean.domain.linkmetadata.entity.vo.Url;
 
 @DataJpaTest
 class LinkMetadataRepositoryTest {
