+++ src/test/java/com/meoguri/linkocean/domain/util/Fixture.java
@@ -3,8 +3,8 @@ package com.meoguri.linkocean.domain.util;
 import static com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType.*;
 
 import com.meoguri.linkocean.domain.bookmark.entity.Bookmark;
-import com.meoguri.linkocean.domain.bookmark.entity.LinkMetadata;
 import com.meoguri.linkocean.domain.bookmark.entity.Tag;
+import com.meoguri.linkocean.domain.linkmetadata.entity.LinkMetadata;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.domain.user.entity.User;
 import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
